+++ packages/cli/src/settings.ts
@@ -14,63 +14,80 @@ import Debug from 'debug';
 
 const debug = Debug('cannon:cli:settings');
 
+/**
+ * Local User Settings for CLI context
+ */
 export type CliSettings = {
-  /// provider used for `build`
-  providerUrl?: string;
+  /**
+   * provider used for `build` defaults to 'frame,direct' https://github.com/floating/eth-provider#presets
+   */
+  providerUrl: string;
 
-  /// private key(s) of default signer that should be used for build, comma separated
+  /**
+   * private key(s) of default signer that should be used for build, comma separated
+   */
   privateKey?: string;
 
-  /// the url of the IPFS endpoint to use as a storage base. defaults to localhost IPFS
+  /**
+   * the url of the IPFS endpoint to use as a storage base. defaults to localhost IPFS
+   */
   ipfsUrl?: string;
 
-  /// the IPFS url to use when publishing. If you have an IPFS cluster, or a pinning service, this is a good place to put its IPFS Proxy publish endpoint. If not specified, your packages wont be uploaded to remote ipfs.
+  /**
+   * the IPFS url to use when publishing. If you have an IPFS cluster, or a pinning service, this is a good place to put its IPFS Proxy publish endpoint. If not specified, your packages wont be uploaded to remote ipfs.
+   */
   publishIpfsUrl?: string;
 
-  /// URL to use to write a package to the registry.
-  registryProviderUrl?: string;
+  /**
+   * URL to use to write a package to the registry. Defaults to `frame,${DEFAULT_REGISTRY_PROVIDER_URL}`
+   */
+  registryProviderUrl: string;
 
-  /// chain Id of the registry. Defaults to `1`. Overridden by `registryProviderUrl`
+  /**
+   * chain Id of the registry. Defaults to `1`. Overridden by `registryProviderUrl`
+   */
   registryChainId: string;
 
-  /// Address of the registry
+  /**
+   * Address of the registry
+   */
   registryAddress: string;
 
-  /// Which registry to read from first. Defaults to `onchain`
+  /**
+   * Which registry to read from first. Defaults to `onchain`
+   */
   registryPriority: 'local' | 'onchain';
 
-  /// Directory to load configurations from, for local registry, and
+  /**
+   * Directory to load configurations from, for local registry, and
+   */
   cannonDirectory: string;
 
-  // URL of etherscan API for verification
+  /**
+   * URL of etherscan API for verification
+   */
   etherscanApiUrl: string;
 
-  // Etherscan API Key for verification
+  /**
+   * Etherscan API Key for verification
+   */
   etherscanApiKey: string;
 
-  // Whether to suppress extra output
+  /**
+   * Whether to suppress extra output
+   */
   quiet: boolean;
 
-  // Gas price to use for transactions
+  /**
+   * Gas price to use for transactions
+   */
   gasPrice?: string;
 
-  // Base and Priority gas fee to use for transactions - EIP1559
+  /**
+   * Base and Priority gas fee to use for transactions - EIP1559
+   */
   gasFee?: string;
   priorityGasFee?: string;
-};
-
-const getRegistryProviderUrl = (fileSettings: any, privateKey: string): string => {
-  const registryProviderUrl = process.env.CANNON_REGISTRY_PROVIDER_URL || fileSettings.registryProviderUrl;
-
-  if (registryProviderUrl && privateKey) {
-    return registryProviderUrl;
-  }
-  if (registryProviderUrl && !privateKey) {
-    console.warn(
-      `Using Frame instead of configured registryProviderUrl (${registryProviderUrl}), supply --private-key to change.`
-    );
-  }
-  return `frame,${DEFAULT_REGISTRY_PROVIDER_URL}`;
 };
 
 // TODO: this function is ugly
@@ -92,15 +109,18 @@ function _resolveCliSettings(overrides: Partial<CliSettings> = {}): CliSettings 
     );
     console.warn(`using default settings (${DEFAULT_REGISTRY_IPFS_ENDPOINT}, ${DEFAULT_REGISTRY_PROVIDER_URL})`);
   }
-  const privateKey = (process.env.CANNON_PRIVATE_KEY || fileSettings.privateKey) as string;
+
   const finalSettings = _.assign(
     {
       cannonDirectory: untildify(process.env.CANNON_DIRECTORY || DEFAULT_CANNON_DIRECTORY),
       providerUrl: process.env.CANNON_PROVIDER_URL || fileSettings.providerUrl || 'frame,direct',
-      privateKey,
+      privateKey: (process.env.CANNON_PRIVATE_KEY || fileSettings.privateKey) as string,
       ipfsUrl: process.env.CANNON_IPFS_URL || fileSettings.ipfsUrl,
       publishIpfsUrl: process.env.CANNON_PUBLISH_IPFS_URL || fileSettings.publishIpfsUrl,
-      registryProviderUrl: getRegistryProviderUrl(fileSettings, privateKey),
+      registryProviderUrl:
+        process.env.CANNON_REGISTRY_PROVIDER_URL ||
+        fileSettings.registryProviderUrl ||
+        `frame,${DEFAULT_REGISTRY_PROVIDER_URL}`,
       registryChainId: process.env.CANNON_REGISTRY_CHAIN_ID || fileSettings.registryChainId || '1',
       registryAddress: process.env.CANNON_REGISTRY_ADDRESS || fileSettings.registryAddress || DEFAULT_REGISTRY_ADDRESS,
       registryPriority: (process.env.CANNON_REGISTRY_PRIORITY ||
@@ -113,7 +133,22 @@ function _resolveCliSettings(overrides: Partial<CliSettings> = {}): CliSettings 
     _.pickBy(overrides)
   );
 
-  debug('got settings', finalSettings);
+  // Filter out private key for logging
+  /* eslint-disable @typescript-eslint/no-unused-vars */
+  const { cannonDirectory, privateKey, etherscanApiKey, ...filteredSettings } = finalSettings;
+
+  // Filters out API keys
+  filteredSettings.providerUrl = filteredSettings.providerUrl
+    ? filteredSettings.providerUrl.replace(RegExp(/[=A-Za-z0-9_-]{32,}/), '*'.repeat(32))
+    : '';
+  filteredSettings.registryProviderUrl = filteredSettings.registryProviderUrl
+    ? filteredSettings.registryProviderUrl!.replace(RegExp(/[=A-Za-z0-9_-]{32,}/), '*'.repeat(32))
+    : '';
+  filteredSettings.publishIpfsUrl = filteredSettings.publishIpfsUrl
+    ? filteredSettings.publishIpfsUrl!.replace(RegExp(/[=AZa-z0-9_-]{32,}/), '*'.repeat(32))
+    : '';
+
+  debug('got settings', filteredSettings);
 
   return finalSettings;
 }