+++ packages/cli/src/types.ts
@@ -3,6 +3,7 @@ export type PackageSettings = { [k: string]: string };
 export interface PackageSpecification {
   name: string;
   version: string;
+  preset?: string;
   settings: PackageSettings;
 }