+++ packages/cli/src/util/params.ts
@@ -1,21 +1,9 @@
 import _ from 'lodash';
 import { InvalidArgumentError } from 'commander';
 import { PackageSpecification, PackageSettings } from '../types';
+import { PackageReference } from '@usecannon/builder';
 
-const packageRegExp = /^(?<name>@?[a-z0-9][a-z0-9-]+[a-z0-9])(?::(?<version>.+))?$/;
 const settingRegExp = /^(?<key>[a-z0-9-_]+)=(?<value>.*)$/i;
-
-export function parsePackageRef(val: string) {
-  const match = val.match(packageRegExp);
-
-  if (!match) {
-    throw new InvalidArgumentError(`Invalid package name "${val}". Should be of the format <package-name>:<version>`);
-  }
-
-  const { name, version = 'latest' } = match.groups!;
-
-  return { name, version };
-}
 
 export function parseInteger(val: string) {
   const parsedValue = Number.parseInt(val);
@@ -45,10 +33,12 @@ export function parseSettings(values: string[] = []) {
 }
 
 export function parsePackageArguments(val: string, result?: PackageSpecification): PackageSpecification {
-  const packageMatch = val.match(packageRegExp);
+  const packageMatch = PackageReference.isValid(val);
 
   if (!result && !packageMatch) {
-    throw new InvalidArgumentError('First argument should be a cannon package name, e.g.: greeter:1.0.0');
+    throw new InvalidArgumentError(
+      'First argument should be a cannon package name, e.g.: greeter:1.0.0 or greeter:latest@main'
+    );
   }
 
   if (result && !_.isEmpty(result) && packageMatch) {
@@ -56,11 +46,18 @@ export function parsePackageArguments(val: string, result?: PackageSpecification
   }
 
   if (packageMatch) {
-    const { name, version = 'latest' } = packageMatch.groups!;
+    const pkg = new PackageReference(val);
+
+    if (!_.isEmpty(result)) {
+      throw new InvalidArgumentError('You can only specify a single cannon package');
+    }
+
+    const { name, version, preset } = pkg;
 
     const def = {
       name,
       version,
+      preset,
       settings: {},
     };
 
@@ -78,13 +75,16 @@ export function parsePackageArguments(val: string, result?: PackageSpecification
 }
 
 export function parsePackagesArguments(val: string, result: PackageSpecification[] = []) {
-  const packageMatch = val.match(packageRegExp);
+  const packageMatch = PackageReference.isValid(val);
+
   if (packageMatch) {
-    const { name, version = 'latest' } = packageMatch.groups!;
+    const pkg = new PackageReference(val);
+    const { name, version, preset } = pkg;
 
     const def = {
       name,
       version,
+      preset,
       settings: {},
     };